Mostly fix handling of implicit parameter dataTypes
The spec and the Pets example state that dataType when applied to
ApiParamsImplicit should cause the specified type to be emitted into the
model portion of the documentation file.  This wasn't happening previously.

The oversight was that there was no code calling
docParam.setValueTypeInternal and therefore when it came time to emit the
models the parameter was ignored.

This change handles dataType for implicit params very much like
responseClass.  This handles _most_ usecases, however, I wasn't sure how
you wanted to handle the parameter analog of isResponseMultiValue.  Would
it be correct to check allowMultiple and treat it similarily (specifying
in the docs that a List[FooObj] is expected).

The reason the Pet example seemed to work, is that in the example
PetController some other method was returning a Pet object, thus forcing
the object into the emitted model section.  However I discovered this bug
because my usecase wasn't so lucky.  ;-)

Thanks for the great tool!