Paths should start with a slash
In the jaxrs version of this code the generated path always starts with
a slash.  Which is a good thing, because swagger-ui concatenates the path
directly with the baseUrl.

In the play2 version it was just starting from the method name (no leading
slash).  This made the swagger-ui 'try it out' button not work with play2
based classes (as seen with the play2 version of the pets sample application).

With this change the samples/scala-play2 will work with swagger-ui
'try it out' (and it now works for my app as well)